fix: rename to AlgebraicSolver for clarity, add accessor for unconditionally getting linear solver instance
