+++ src/solvers/thermal_operators.cpp
@@ -22,7 +22,7 @@ DynamicConductionOperator::DynamicConductionOperator(std::shared_ptr<mfem::ParFi
       old_dt_(-1.0)
 {
   // Set the mass solver options (CG and Jacobi for now)
-  M_solver_ = SystemSolver(fespace_->GetComm(), params);
+  M_solver_ = AlgebraicSolver(fespace_->GetComm(), params);
 
   M_solver_.solver().iterative_mode = false;
   auto M_prec                       = std::make_unique<mfem::HypreSmoother>();
@@ -30,7 +30,7 @@ DynamicConductionOperator::DynamicConductionOperator(std::shared_ptr<mfem::ParFi
   M_solver_.SetPreconditioner(std::move(M_prec));
 
   // Use the same options for the T (= M + dt K) solver
-  T_solver_ = SystemSolver(fespace_->GetComm(), params);
+  T_solver_ = AlgebraicSolver(fespace_->GetComm(), params);
 
   T_solver_.solver().iterative_mode = false;
